feat: check forbidden for deleteBlogUseCase

+++ src/features/blog/application/useCases/delete.blog.use.case.ts
@@ -1,9 +1,14 @@
 import { CommandHandler } from '@nestjs/cqrs';
 import { BlogRepositoryTypeorm } from '@blog/repository/typeorm/blog.repository.typeorm';
 import { BlogQueryRepositoryTypeorm } from '@blog/repository/typeorm/blog.query.repository.typeorm';
+import { userInfo } from 'os';
+import { ForbiddenException } from '@nestjs/common';
 
 export class DeleteBlogCommand {
-  constructor(public id: string) {}
+  constructor(
+    public id: string,
+    public userId: string
+  ) { }
 }
 
 @CommandHandler(DeleteBlogCommand)
@@ -11,13 +16,14 @@ export class DeleteBlogUseCase {
   constructor(
     private blogRepository: BlogRepositoryTypeorm,
     private blogQueryRepository: BlogQueryRepositoryTypeorm,
-  ) {}
+  ) { }
 
   async execute(command: DeleteBlogCommand): Promise<boolean> {
-    const { id } = command;
+    const { id, userId } = command;
 
-    const blog = await this.blogQueryRepository.findBlogViewById(id);
+    const blog = await this.blogQueryRepository.findFullBlogById(id);
     if (!blog) return false;
+    if (blog.user.id !== userId) throw new ForbiddenException()
 
     const isDelete = await this.blogRepository.deleteBlogById(blog.id);
     if (!isDelete) return false;
